WIP store session (1)
