--- src/main/java/org/nmfw/foodietree/domain/store/controller/StoreSignOutController.java
@@ -1,26 +1,0 @@
-package org.nmfw.foodietree.domain.store.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import javax.servlet.http.HttpSession;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-@RequestMapping("/store")
-public class StoreSignOutController {
-    @GetMapping("/sign-out")
-    public String signOut(HttpSession session){
-
-        //세션구하기
-        //HttpSession session = request.getSession();
-
-        // 세션에서 로그인 기록 삭제
-        session.removeAttribute("login");
-
-        // 세션을 초기화 (reset)
-        session.invalidate();
-
-        // 홈으로 보내기
-        return "redirect:/";
-    }
-}
