+++ src/main/java/org/nmfw/foodietree/domain/store/mapper/StoreMapper.java
@@ -1,6 +1,7 @@
 package org.nmfw.foodietree.domain.store.mapper;
 
 import org.apache.ibatis.annotations.Mapper;
+import org.nmfw.foodietree.domain.customer.dto.request.AutoLoginDto;
 import org.nmfw.foodietree.domain.store.entity.Store;
 
 @Mapper
@@ -13,4 +14,6 @@ public interface StoreMapper {
     Store findOne(String storeId);
     
     boolean storeSave(Store store);
+
+    void updateAutoLogin(AutoLoginDto dto);
 }
