+++ src/main/resources/mappers/StoreMapper.xml
@@ -24,4 +24,10 @@
       (store_id, password)
     VALUES (#{storeId}, #{password})
   </insert>
+
+  <update id="updateAutoLogin">
+    UPDATE tbl_store
+    SET session_id = #{sessionId}, limit_time = #{limitTime}
+    WHERE customer_id = #{id}
+  </update>
 </mapper>
